refactor : layout components
