+++ src/layout/main/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import '@scss/layout/MainLayout.scss';
+import './index.scss';
 import Header from '@module/Header';
 import Footer from '@module/Footer';
 import ScrollProgress from '@components/ui/ScrollProgress';
