+++ src/layout/side/index.scss
@@ -1,6 +1,6 @@
 @import '@scss/abstracts/mixin';
 
-.sidebar-layout {
+.side-layout {
   position: fixed;
   top: 0;
   left: 0;
