+++ src/pages/list.tsx
@@ -4,7 +4,7 @@ import { HeadFC, PageProps, graphql } from 'gatsby';
 import '@scss/global.scss';
 import '@scss/pages/PostsPage.scss';
 import { getNotionNodeAll } from '@services/use-notion';
-import MainLayout from '@layout/MainLayout';
+import MainLayout from '@layout/main';
 import { NotionContext } from '@store/rootStore';
 import { INotionContext, NotionNode } from '@types';
 import { classifyPost, getParseListByNodes } from '@utils/notionUtils';
